+++ lib/app_localzations.dart
@@ -3,8 +3,6 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:medmap/main.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class AppLocalizations {
   final Locale locale;
@@ -18,30 +16,14 @@ class AppLocalizations {
     return Localizations.of<AppLocalizations>(context, AppLocalizations);
   }
 
-  void keepLocalKey(String localkey) async {
-    final _prefs = await SharedPreferences.getInstance();
-    await _prefs.remove('language_code');
-    await _prefs.setString('language_code', localkey);
-  }
-
-  Future<String> readLocalKey() async {
-    final _prefs = await SharedPreferences.getInstance();
-    return _prefs.getString('language_code') ?? 'en';
-  }
-
-  void setLocale(BuildContext context, Locale localkey) async {
-    keepLocalKey(locale.languageCode);
-    MyApp.setLocale(context, locale);
-  }
-
   // Static member to have a simple access to the delegate from the MaterialApp
   static const LocalizationsDelegate<AppLocalizations> delegate =
-      _AppLocalizationsDelegate();
+  _AppLocalizationsDelegate();
 
   Future<bool> load() async {
     // Load the language JSON file from the "lang" folder
-    String jsonString = await rootBundle
-        .loadString('assets/translations/${locale.languageCode}.json');
+    String jsonString =
+    await rootBundle.loadString('assets/translations/${locale.languageCode}.json');
     Map<String, dynamic> jsonMap = json.decode(jsonString);
 
     _localizedStrings = jsonMap.map((key, value) {