+++ lib/views/submenu.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:medmap/const.dart';
 
 class Submenu extends StatelessWidget {
   @override
@@ -9,25 +10,34 @@ class Submenu extends StatelessWidget {
         appBar: AppBar(
           leading: IconButton(
             icon: Icon(Icons.arrow_back), // Back button icon
-            onPressed: () => Navigator.of(context).pop(), // Navigate back when pressed
+            onPressed: () =>
+                Navigator.of(context).pop(), // Navigate back when pressed
           ),
           actions: [
-            Row(
-              children: [
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    SizedBox(height: 10),
-                    Text('Profile Name', style: TextStyle(fontSize: 16)), // User's name
-                    Text('Company Name', style: TextStyle(fontSize: 14, color: Colors.grey[600])), // User's role
-                  ],
-                ),
-                SizedBox(width: 10), // Space between the avatar and the text
-                CircleAvatar(
-                  backgroundImage: NetworkImage('https://example.com/profile_picture.jpg'), // Replace with your profile picture URL
-                  radius: 30, // Adjust the size of the circle avatar
-                ),
-              ],
+            Padding(
+              padding: EdgeInsets.all(5.0), // Adjust the value as needed
+              child: Row(
+                children: [
+                  Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      // SizedBox(height: 10),
+                      Text('Profile Name',
+                          style: TextStyle(fontSize: 16)), // User's name
+                      Text('Company Name',
+                          style: TextStyle(
+                              fontSize: 14,
+                              color: Colors.grey[600])), // User's role
+                    ],
+                  ),
+                  SizedBox(width: 10), // Space between the avatar and the text
+                  CircleAvatar(
+                    backgroundImage: NetworkImage(
+                        'https://example.com/profile_picture.jpg'), // Replace with your profile picture URL
+                    radius: 30, // Adjust the size of the circle avatar
+                  ),
+                ],
+              ),
             ),
           ],
         ),
@@ -55,13 +65,27 @@ class CardsLayout extends StatelessWidget {
             color: Colors.white, // Ensure the card itself is also white
             elevation: 0, // Remove shadow
             margin: EdgeInsets.symmetric(vertical: 20, horizontal: 20),
-            child: Center( // Wrap the Column with Center to center its contents
+            child: Center(
+              // Wrap the Column with Center to center its contents
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  Icon(Icons.image, size: 50),
+                  Image.asset(
+                    Const.submenu_report,
+                    fit: BoxFit.contain,
+                    height: 30,
+                  ),
+                  // Icon(Icons.image, size: 50),
                   SizedBox(height: 10),
-                  Text('Request Market Study Report', style: TextStyle(fontSize: 10)),
+                  Padding(
+                    padding: EdgeInsets.symmetric(
+                        horizontal: 5.0), // Add padding to the left and right
+                    child: Text(
+                      'Request Market\nStudy Report',
+                      textAlign: TextAlign.center, // Center the text
+                      style: TextStyle(fontSize: 10),
+                    ),
+                  ),
                 ],
               ),
             ),
@@ -74,13 +98,23 @@ class CardsLayout extends StatelessWidget {
             color: Colors.white,
             elevation: 0,
             margin: EdgeInsets.symmetric(vertical: 20, horizontal: 20),
-            child: Center( // Wrap the Column with Center to center its contents
+            child: Center(
+              // Wrap the Column with Center to center its contents
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  Icon(Icons.image, size: 50),
+                  Image.asset(
+                    Const.submenu_event,
+                    fit: BoxFit.contain,
+                    height: 30,
+                  ),
                   SizedBox(height: 10),
-                  Text('Request for Product Launch Event', style: TextStyle(fontSize: 10)),
+                  Padding(
+                    padding: EdgeInsets.symmetric(horizontal: 5.0),
+                    child: Text('Request for Product Launch Event',
+                        textAlign: TextAlign.center,
+                        style: TextStyle(fontSize: 10)),
+                  ),
                 ],
               ),
             ),
@@ -93,13 +127,24 @@ class CardsLayout extends StatelessWidget {
             color: Colors.white,
             elevation: 0,
             margin: EdgeInsets.symmetric(vertical: 20, horizontal: 20),
-            child: Center( // Wrap the Column with Center to center its contents
+            child: Center(
+              // Wrap the Column with Center to center its contents
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  Icon(Icons.image, size: 50),
+                  Image.asset(
+                    Const.submenu_design,
+                    fit: BoxFit.contain,
+                    height: 30,
+                  ),
                   SizedBox(height: 10),
-                  Text('Request Product Brochure Design', style: TextStyle(fontSize: 10)),
+                  Padding(
+                    padding: EdgeInsets.symmetric(
+                        horizontal: 5.0), // Add padding to the left and right
+                    child: Text('Request Product Brochure Design',
+                        textAlign: TextAlign.center,
+                        style: TextStyle(fontSize: 10)),
+                  ),
                 ],
               ),
             ),
@@ -112,13 +157,24 @@ class CardsLayout extends StatelessWidget {
             color: Colors.white,
             elevation: 0,
             margin: EdgeInsets.symmetric(vertical: 20, horizontal: 20),
-            child: Center( // Wrap the Column with Center to center its contents
+            child: Center(
+              // Wrap the Column with Center to center its contents
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  Icon(Icons.image, size: 50),
+                  Image.asset(
+                    Const.submenu_privacy,
+                    fit: BoxFit.contain,
+                    height: 30,
+                  ),
                   SizedBox(height: 10),
-                  Text('Privacy Policy', style: TextStyle(fontSize: 10)),
+                  Padding(
+                    padding: EdgeInsets.symmetric(
+                        horizontal: 5.0), // Add padding to the left and right
+                    child: Text('Privacy Policy',
+                        textAlign: TextAlign.center,
+                        style: TextStyle(fontSize: 10)),
+                  ),
                 ],
               ),
             ),